Added Orders and Modified UI

+++ backend/Ecommerce.API/DTOs/DashboardSummaryDto.cs
@@ -1,7 +1,7 @@
 ﻿namespace Ecommerce.API.DTOs;
 public class DashboardSummaryDto
 {
-    public int TotalSales { get; set; }
+    public decimal TotalSales { get; set; }
     public int TotalOrders { get; set; }
     public int TotalProducts { get; set; }
     public int TotalCustomers { get; set; }
